// src/app/models/admin/dangkyhoatdongkhoangsan/dangkytralai/dkkhaithactralai.model.ts
export class OutputDkKhaiThacTraLaiModel {
    iddangkykhaithac: string;
    dientichdacapphep: number;
    dientichkhaithac: number;
    dientichtralai: number;
    donvidientich: string;
    lydotralai: string;
    geom: string;
    geovn: string;
    geowgs: string;
    hequychieu: string;
    idhoso: string;
    idgiayphep: string;
    loaitralai: number;
    sogiayphep: string;
    ngaycapphep: number;
    siteid: string;
}

export class InputDkKhaiThacTraLaiModel {
    iddangkykhaithac: string;
    dientichdacapphep: number;
    dientichkhaithac: number;
    dientichtralai: number;
    donvidientich: string;
    lydotralai: string;
    hequychieu: string;
    idhoso: string;
    idgiayphep: string;
    loaitralai: number;
}